+++ lib/zaikio/procurement.rb
@@ -9,21 +9,21 @@ require "zaikio/procurement/base"
 require "zaikio/procurement/address"
 require "zaikio/procurement/article"
 require "zaikio/procurement/variant"
-require "zaikio/procurement/substrate_search"
+require "zaikio/procurement/variant_search"
 require "zaikio/procurement/sku"
-require "zaikio/procurement/price"
 require "zaikio/procurement/supplier"
-require "zaikio/procurement/contract"
 require "zaikio/procurement/contract_request"
-require "zaikio/procurement/sales_group"
-require "zaikio/procurement/sales_group_membership"
 require "zaikio/procurement/order"
 require "zaikio/procurement/order_line_item"
 require "zaikio/procurement/delivery"
-require "zaikio/procurement/delivery_line_item"
-require "zaikio/procurement/current_supplier"
 require "zaikio/procurement/material_requirement"
-require "zaikio/procurement/order_conditions_inquiry"
+require "zaikio/procurement/material"
+require "zaikio/procurement/pricing"
+require "zaikio/procurement/site"
+require "zaikio/procurement/availability"
+require "zaikio/procurement/quantity"
+require "zaikio/procurement/job"
+require "zaikio/procurement/person"
 
 module Zaikio
   module Procurement
@@ -48,12 +48,8 @@ module Zaikio
         AuthorizationMiddleware.token = original_token
       end
 
-      def connection_path
-        "#{configuration.host}/#{configuration.flavor.to_s.pluralize}/api/v1/"
-      end
-
       def create_connection
-        self.connection = Faraday.new(url: connection_path,
+        self.connection = Faraday.new(url: configuration.host,
                                       ssl: { verify: configuration.environment != :test }) do |c|
           c.request     :json
           c.response    :logger, configuration&.logger, headers: false