+++ lib/zaikio/procurement/configuration.rb
@@ -4,22 +4,19 @@ module Zaikio
   module Procurement
     class Configuration
       HOSTS = {
-        development: "https://procurement.zaikio.test",
-        test: "https://procurement.zaikio.test",
-        staging: "https://procurement.staging.zaikio.com",
-        sandbox: "https://procurement.sandbox.zaikio.com",
-        production: "https://procurement.zaikio.com"
+        development: "https://procurement.zaikio.test/api/v2/",
+        test: "https://procurement.zaikio.test/api/v2/",
+        staging: "https://procurement.staging.zaikio.com/api/v2/",
+        sandbox: "https://procurement.sandbox.zaikio.com/api/v2/",
+        production: "https://procurement.zaikio.com/api/v2/"
       }.freeze
 
-      FLAVORS = %i[consumer supplier].freeze
-
       attr_accessor :host
-      attr_reader :environment, :flavor
+      attr_reader :environment
       attr_writer :logger
 
       def initialize
         @environment = :sandbox
-        @flavor      = :consumer
       end
 
       def logger
@@ -29,12 +26,6 @@ module Zaikio
       def environment=(env)
         @environment = env.to_sym
         @host = host_for(environment)
-      end
-
-      def flavor=(flavor)
-        raise StandardError.new, "Invalid Zaikio::Procurement flavor '#{flavor}'" unless FLAVORS.include?(flavor)
-
-        @flavor = flavor
       end
 
       private