+++ lib/zaikio/procurement/job.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module Procurement
+    class Job < Base
+      # Attributes
+      attributes :anticipated_costs, :client, :description, :link, :reference
+    end
+  end
+end