+++ lib/zaikio/procurement/order_line_item.rb
@@ -1,16 +1,12 @@
 module Zaikio
   module Procurement
     class OrderLineItem < Base
-      uri "order_line_items(/:id)"
-      include_root_in_json :order_line_item
-
       # Attributes
-      attributes :amount, :amount_in_base_unit, :base_unit, :catalog_price, :confirmed_price, :description,
-                 :order_number, :tax_rate, :taxes, :total_gross_price, :total_net_price, :unit, :order_id,
-                 :sku_id, :created_at, :updated_at
+      attributes :references, :amount, :amount_in_base_unit, :base_unit, :confirmed_delivery_at, :description,
+                 :order_number, :created_at, :updated_at
 
       # Associations
-      belongs_to :order, class_name: "Zaikio::Procurement::Order",
+      has_one :pricing,  class_name: "Zaikio::Procurement::Pricing",
                          uri: nil
       has_one :sku,      class_name: "Zaikio::Procurement::Sku",
                          uri: nil