+++ lib/zaikio/procurement/person.rb
@@ -1,0 +1,8 @@
+module Zaikio
+  module Procurement
+    class Person < Base
+      # Attributes
+      attributes :full_name
+    end
+  end
+end