+++ lib/zaikio/procurement/pricing.rb
@@ -1,0 +1,6 @@
+module Zaikio
+  module Procurement
+    class Pricing < Base
+    end
+  end
+end